:recycle: Refactor code to make Use Cases smarter
Updated the code handle all the RX schedulers on the Domain layer,
making it less complex to be called from the Feature modules.
DAOs updated to provide a Completable in the no return calls, making it
easier to standardize the Domain module.
Also, the Room dependencies where updated to match the Rx new
implementation