:arrow_up: Update all V2 related dependencies
Compose, Gradle and Kotlin dependencies updated in order to keep the
development.

At the moment is very difficult to the project compile and keep passing
in the CI, so this commit will go directly to V2 branch. When Google
fix their issues, I can proceed and make the CI pass again.