:sparkles: Update Task Detail to be more static
Changing of plans: when this screen was first created, the idea was to
make it very dynamic; if the database is updated while the user is on
screen, then the updated content will be shown automatically. The
problem is that this conflicts a little with my current idea of the
screen: all data is saved as soon (with debounce) as the user inputs it.
So this created a problem where the screen was been recomposed multiple
times while the user was inputting data, which leads to a bad
experience.
The solution at the moment is to make the screen more static: it will
continue to save as the user inputs, but external changes will not
affect the screen on fly. Basically it was updated from "Flow" to a
regular one-time call Coroutine. In the future this decision can be
revisited once more knowledge about Compose is developed.