♻️ Refactor TaskDetail to load information from two ViewModels
Now the TaskDetail loads information from TaskDetailViewModel and
TaskCategoryViewModel in order to better separate the concerns about
both states. Due to this change, the states needs to be propagated to
the views below and some additional handling was needed, specially in
the CategorySelection.

As a bonus, based on the conversation with @adamp, replaced the
`StateFlow` to `State` to avoid unnecessary complexity between their
conversion.