✨ Create AlarmSelection component
New component to select the alarm date and time no be notified and if
the alarm is repeating with the selected interval.
New tests added to validate the component behavior and once there is no
Date and Time Dialog support in Compose yet, the default from Android
View System is used. Because of this, Espresso was re-add in the
project.