Category screen is showing wrong info when rotating screen
[root-cause] For some reason that I do not have knowledge yet, the
rememberSaveable is not behaving correctly. Sometimes, during
orientation changes, it remembers the data from one interaction before.
This could be an issue in the Compose framework after all (see link
above)

[solution] At the moment, the code was updated to use a simple remember.
The only "issue" is that if the user updates the data and rotates the
screen, it will clean their input and load the original data again.

[see] https://issuetracker.google.com/issues/182403380