✅ Remove flakiness from AlarmSelectionTest (#157)
Add the Test Rule to disable animations when using components from
Android View System. Also a sleep was added in order to wait for the
Android View dialog load. For sure it is not the best approach, however
when a proper Compose dialog is launched, this problem will not happen.