✨ Update the code to show the permission dialog when needed
Code updated to hoist the verification down to the UI and showing the
permission dialog if needed. New string were added and translated to
this new feature. A new Instrumented Test was added to validate the
expected behavior.

Also the AndroidX Test Core was updated due to weird issues. For more
information check: https://issuetracker.google.com/issues/194419355