🚨 Fix broken tests
Okay, we are having again issues related to logs on Kotlin-only modules.
By now, in order to fix the tests as soon as possible, the logs in
:domain are commented.

In a next pull request I will implement different loggers for each
module type. This may fix this issue for once and all.