✨ Create JVM Convention Plugins to replace KTS plugins
New JVM Convention plugins created for the common module usages, such
as Compose, Dynamic Delivery, Android and Kotlin. Also implemented the
ones for Quality.

Some unsafe calls were needed for libraries accessed from the JVM
Convention files, which is not possible to be accessed safely, as
acknowledge in this discussion:
https://github.com/gradle/gradle/issues/15383