💩 Make `test_updateCategoryColor` flaky
That's not a thing I would like to do, but I'm stuck with this issues
for months. A Issue Tracker was opened for support, but in the meantime
I will have this test annotated as Flaky and trying it 5 times before
actually failing the tests.
Let's see if now I'm able to finally merge this change.