♻️ Refactor AlarmPermission to include the new permission
The AlarmPermission interface already contained the permission to send
exact alarms, which is a special permission already granted but can be
removed by the user.

Once the new notification permission today for Alkaa is only related to
Alarm, it was added on the same interface. If needed (for example,
having other notification across the app) it could be refactored to a
separate interface (and even different module).