💄 Update Composable to check for permission before setting an alarm
The Composable for AlarmSelection was updated to check for the
permission when the user clicks to set a new alarm. A very simple
logic was created along with the previous exact alarm permission. In the
worst case scenario, the user may need to see and grant two dialog
permissions in a row. However, this is unlikely since the exact alarm
permission is already granted by default.

In order to avoid the size of the file increase a lot, the dialogs were
moved to a different file. Lastly, new strings were added to inform the
user about the permissions both in English and Portuguese.