✨ Improve the permission flow experience
Now the permission flow allows the user to see a rationale dialog
explaining the importance of the permission, when they previously
denied it. It will show a dialog with information and a call to action
redirecting to the app settings to manually grant it.

Some refactor was also made in the AlarmSelection due to its complexity.
This was made by creating a state holder with all relevant data to be
abstract from the Composable and also allowing better readability.

Lastly, more strings were added in both English and Portuguese to help
in the flow.

Special thanks to @GabrielBottari for the tips.