♻️ Hoist the Permission Notification check
After some initial Instrumented Tests, I noticed that in devices API <=
33 the `rememberPermissionState` is returning `PermissionStatus.Denied`.
That makes sense since the official `checkSelfPermission` API returns
`-1` in that case.

Based on that, we can't rely exclusively on the Accompanist Permission
API to remember the state. The proposed solution is hoisting the API
verification to the proper layer and mock a fake
`PermissionStatus.Granted` if the device does not support this kind of
permission.

I opened an issue at the Accompanist Permission sharing this solutions
and some suggestions: https://github.com/google/accompanist/issues/1328