⏪️ Revert to Gradle Managed Devices (yes, again)
The problem was not related to the Android Gradle Managed Devices after
all. It seems that my new `FlakyTest()` (which was created because some
Compose tests were flaky on CI) made the tests EVEN FLAKIER. That's
ironic. Or not, because the name says that they are flaky... Anyway,
reverting it back.