♻️ Update the Glance implementation
In the recent versions of Glance API, relying only in a Coroutine to
update the value is no longer enough. During my tests, the Alkaa process
 might be dead or the Coroutine cancelled in the middle of the
 execution, which causes the Widget to show outdated data or even empty
 list.

The proposed solution uses WorkManger to so the heavy processing and
DataStore to make the data quick and easy accessible for the Widget,
regardless of app state. This implementations is heavily based on the
Google's code sample o GitHub:
https://github.com/android/user-interface-samples/tree/main/AppWidget