✨ Implement a new CoroutineScope to be injected in the ViewModels
The Glance feature exposed some issues with the current ViewModelScope
on Alkaa: as soon as the Composable is gone, the ViewModel dies and the
execution is cancelled. This impacts all the Interactors that are
called after IO operation.

Based on Google's article and some discussion the solution was to create
a new CoroutineScope and inject in the following case:

> When the suspend function must be executed regardless of which screen
the user is, use the injected scope. If the execution should be
cancelled, then use viewModelScope.