👷 Force Check to be dependent on Lint
For some reason, this relation between Check and Lint should be already
created by default. I'm not sure if this is not working due to all my
custom setup, especially on `com.escodro.kotlin-quality.gradle` where it
explicitly sets the dependencies. Anyway, the statement to force this
relation is now added and Lint is working fine again.