🐞 Fix issue when sometimes the app does not add or update tasks (#401)
In the past, we migrated some of the ViewModelScopes to ApplicationScope
to ensure that the execution would run regardless of the view state.
However, in a very specific flow I was cancelling this Scope.

This cancellation was happening when the Glance widget was removed from
the home screen. When this cancellation was called, than no new
execution was made in that scope.

The quick fix for now is simply removing the cancellation, however I
will investigate how to avoid this issue in the near feature by creating
some kind of protection to its cancellation.