♻️ Create custom CoroutineScope to be used application-wide
A new class called `AppCoroutineScope` was created to wrap the
functionality of a `CoroutineScope` without exposing the `.cancel()`
API. Since this scope should match the lifecycle of the App, it should
never be cancelled. This class will avoid human error by not exposing
the function to cancel the scope.

The decision to create a wrapper instead of implementing the
`CoroutineScope` interface was because the `.cancel()` function is not
part of the interface itself, it's an extension function. Due to the
Liskov Substitution Principle, it would be possible to attribute a
`CustomCoroutineScope` in a `CoroutineScope` interface, making it
possible to access the `.cancel()` function even if it was implemented
in the custom one.

For the moment, only the `.launch()` function is exposed. If other
functions (such as `.async()`) is needed, than future changes will add
them.