🔥 Remove the weird BowWithConstraints from TaskList
In the original concept, this BowWithConstraints would show two columns
when the app is on landscape or bigger screens. However, it does not
work very well.
When the keyboard opens it show two columns, creating a very bad UX.

This is gone for now and will be improved in future commits.