Update Compose and Kotlin-related dependencies (#445)
* Update compose to v1.4.0

* 🚨 Fix CI

Compose 1.4.0 introduced new functions that were manually added in the
code. The new functions are now used and the required opt-in added.

* ⬆️ Update Compose Compiler and Kotlin-related dependencies

Compose Compiler, Kotlin, KSP and Material Icons versions upgraded

* 🚨 Fix annoying Java validation error

When updating the Compiler and Kotlin version, a new error appeared
about Java versions not matching. The issue stated that Java 1.8 and 11
were conflicting, even though there is no Java 1.8 direct reference in
the project. The solution is to add a new parameter in the
`gradle.properties` file.

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>