🐞 Fix Bottom Nav colors not updating after theme change
When updating the system theme under the Preference screen, the colors
in the Bottom Nav would remain the previous one until the app
recomposes. It's weird because the theme change is trigger the high
level recomposition and updating all subsequent components.

The solution for this issue was to remove the custom colors from
`NavigationBarItem` and use the default ones. Now that this component is
Material3 it will get the correct colors without additional setup. This
fix both the issue and the UI that sometimes was not beautiful because
of weird colors.