🔥 Remove CocoaPods
When I was trying to integrate iOS and KMP, several issues were faced. I
tried, naively, to add CocoaPods since some of the samples also used
and it worked. Today is clear that it worked due to _other_ reasons,
not CocoaPods, since it's not needed due to
`embedAndSignAppleFrameworkForXcode` Gradle task. Less code is more
code.