✨ Implement the Home module
A new feature module was created to hold the "Home" UI code. In the
previous structure, this code was inside Android's `:app` but in order
to make it multiplatform, we need a dedicated module. We could put the
logic on `shared`, but my goal to that module is just to be a "umbrella"
with small setup.