✨ Implement the Resources module
In the previous structure, each module would have its own string
directory and localizations. However, the way that Moko is set up, this
is more difficult and will require constant changes in the iOS project
to compile for every single module. The alternative implemented here is
creating a single "Resources" module that holds all the string and every
UI module depends on it.