♻️ Convert the navigation module to KMP and add Voyager library
The existing navigation module was revamped to be a KMP module and
starting using Voyager as the navigation framework. For now, only the
Task and Category flows were added. Future changes will add the
remaining ones and also we will investigate if it's worth it to move the
tab navigation to this library too.

Since the module was an Android library and still referenced, everything
was commented at this moment.