🚧 Disable all the Interactors
In order to make sure that iOS will keeping running, all the Interactors
that are not implemented will be disabled. In the future, a more
scalable solution will be implemented to handle both the cases where
iOS will have a dedicated implementation and also the ones that should
be ignored.