🚧 Fix the unit tests
Temporary solution to fix domain the unit tests. Since the Notification,
Alarm and Widget/Glance features are not ported to KMM, the DI will
provide a nullable injection for those Interactors. Unfortunately, Koin
does not support `getOrNull` with the Constructor DSL so we are back
to the ugly injection temporarily.