✨ Introduce category pre-population on iOS
Since the SQLDelight does have a callback when the database schema is
created, the solution was checking if the database is created before
inserting the categories. The library creates the database on a specific
folder, which is now checked on iOS variant.

This change also required a few structure changes, such as the place of
the strings and implementing the resource access on each platform.