✨ Re-add Tracker feature on Android
The Tracker feature is a Dynamic Feature (on-demand), which is only
supported by Android right now. Due to the nature of this feature the
dependency is inverted, meaning that the `app/shared` module doesn't
know about the feature but the feature knows about the `app/shared`.

Future commits will try to implement this a regular feature for iOS.