🐞 Fix BottomSheet + Keyboard issues
The BottomSheet + Keyboard is now working better on both platforms. On
Android, since the addition of `enableEdgeToEdge()`, the keyboard was
overlapping the bottom sheet. Also, on both platforms, the keyboard was
not being dismissed when the bottom sheet is hidden.

For iOS, unfortunately we don't have a `imePadding()` modifier yet to
handle that. A third-party library was added, but didn't work as
expected. For now, the iOS Keyboard and bottom sheet behavior will
remain the default one.