✨ Re-enable the ability of navigating using deep links
Voyager does not have support for deep links or navigation from outside
of the app at this moment. Because of that, a very simple structure was
created to allow navigation through notifications and widgets using deep
links. For now, it only applies to the Android platform.

The idea is to have a single deep link redirecting to the MainActivity
and send a single extra, which is a sealed interface representing the
destination and the parameters needed. The activity gets this
information from the Intent and passes it to the navigation graph. After
that, all the execution is multiplatform.

The navigation graph consumes the action and sends it to the appropriate
screen (for now, only Task Details). If no information is provided, it
stays on the home screen and no extra steps are taken.

For a more complex deep link interaction, we could send a Uri-like
structure to be handled instead. But since Alkaa does not have any
external deep link provider and we create all the intents internally,
this is simpler and we don't need to worry about type safety.