✨ Implement a composable refresher in the task list
Since we are using more sensitive relative time strings, they will be
outdated if the screen is not updated (e.g. "4 minutes ago"). This new
effect refreshes the jey every minute and recomposes the data fetch. As
this is an UI-related change, the Composable is the one responsible, not
the data flow itself.