🔥 Remove "core" module
Core was a very nice module. At first, it held all the fun extension
functions for the Android application. However, core was greed. It
wanted to hold a lot of common logic, not allowing cleaner and more
focused modules to have their place in the project. Each day, it was
harder and harder to get rid of core. One day, Alkaa decided to go to
the realm of multiplatform, and of course, core didn't like the idea.
Core tried very hard to convince everyone that this was a bad idea, that
this was just a fling and, incredibly, even suggested... Flutter! But
it was too late, and all the other sibling modules were excited to be
converted to KMP and have different source sets. Core was losing its
power since it was impossible to migrate it, and it lost function after
function until it became useless... Today, it's the day we say
"goodbye" to core. And I hope that you find peace in the realm of the
deleted code.