🐞 Fix Android crash when the app Activity goes to background
The app was crashing when the Activity was no longer in focus. This
happened because all the property in the `Screen` objects needs to be
Serializable for correct state management. Since the `WindowSizeClass`
doesn't have this support and all the constructors are private, some
back-and-forth would be needed to make it work properly. For now, the
solution was not holding this parameter in the class and use it directly
in the `Home` Composable.