🐞 Fix bug where the notification does not have the updated title in iOS
In iOS, the notification is scheduled with the task data when it is
created. This means that if the title changes over time, the
notification would show the outdated title when the notification is
launched.

In this proposed fix, the notification is updated when the title
changes, providing the latest info to the interactor.