🔂 Replace the TaskAlarmViewModel logic with the new use case
Now that we have a use case that handles the alarm update, the logic and
tests were simplified in the Ui layer. I don't love the solution to
have to load the task before sending a copy to be modified, but for
now, that's not going to impact the code too much and don't require a
big refactor.