🧃 Update SQLDelight to run the insert + select in a transaction
To ensure that the "select last row" will work properly after inserting
the task, the code was updated to run inside a transaction. This is
required because it "only works on the connection where the write
happened". Despite working fine in Android, this was not working on iOS.

Source: https://github.com/cashapp/sqldelight/issues/1828