🪳 Remove the "refresh task screen" `LaunchedEffect` to allow iOS tests
After spending hours trying to understand why the iOS tests were not
running and getting stuck, I founded out that the issue is the
`rememberRefreshKey()`. In the iOS tests, the `delay` calls are not
being respected and since we have a 1-minute delay for the Task Screen,
it is infinite looping. For now, we are simply removing it.

A new issue was opened in the Compose Multiplatform project to get a
better understanding of the behavior.
https://github.com/JetBrains/compose-multiplatform/issues/4805