🎭 Update GitHub Actions to run the Instrumented tests
Changes made in the build.yml and instrumented_tests.yml to run the
instrumented tests for both iOS and Android. The build.yml was updated
to remove the unit tests from the build step, otherwise, it will try to
run the UI tests during that step. Not the perfect solution, but good
enough for now.

The Android Emulator Runner was also updated and it's running much
faster with the new recommended setup.