✅ Fix tests on iOS
After further investigation, I found out that the iOS tests doesn't have
`LocalLifecycleOwner` by design, apparently since the "actual"
implementation simply throws an exception. The solution was to create a
fake `LifecycleOwner` to be used on the tests.

In addition, a test in `CategoryFlowTest` was update to not be flaky on
iOS.