🐞 Fix Top App Bar title defaulting to "Tasks"
Previously, the code was defaulting to Task if the destination was not
found as a TopLevel. This creates problems when the Top App Bar is
visible and a non top-level is the current destination. In these cases,
it would show the "About" screen with the "Tasks" title.

To fix this the find will return null if it's not a TopLevel and
`filterNotNull` was added to ensure there will be no emission on this
case, leaving the current title active.