⬆️ Update TopLevel navigation detection
There was a bug when switching TopLevel destinations with saved
backstack. Since the `currentTopDestination` was only returning the
current entry, it was not possible to get the TopLevel where this
destination was started, making the Bottom Bar icon not updating
correctly.

The top level destination is now obtained from the backstack, instead of
only from the current entry. This allows the app to properly detect the
correct top level screen when navigating back.