👨‍❤️‍👨 Decouple the permission logic from the Moko library
The `libraries:permission` module was refactored to decouple the main
logic to request and verify the permission from the Moko library. Under
the hood, Moko is still used to handle the iOS and Android permissions,
however, this new custom implementation will allow us to target new
platforms since Moko Permissions only targets the mobile ones for KMP.
This change will also help the port to another library, if needed.