📦 Implement Common Parcelable in KMP
The `kotlin-parcelize` plugin is added to the module for Android.
The `CommonParcelize` annotation was added to enable the Parcelable
generation in KMP, allowing to use the same annotation in the common
code. The `CommonParcelable` interface was created to allow type
aliasing for the different platforms. `CommonIgnoredOnParcel` is the
annotation to ignore properties.